+++ etl/publish/publish_stock_profiles.py
@@ -3,7 +3,9 @@ import pandas as pd
 import numpy as np
 from database.utils import connect_to_db, insert_records, execute_query
 from utils import delete_published_records
+import os 
 
+app_env = os.getenv("APP_ENV", "local")
 
 
 def read_records():
@@ -25,12 +27,15 @@ def main():
     Main function to orchestrate the ETL process for stock profiles.
     """
     conn_local = connect_to_db("localhost")
-    conn_supabase = connect_to_db("supabase")
+    conn_supabase = True
+    if app_env == "local":
+        conn_supabase = connect_to_db("supabase")
     if conn_local and conn_supabase:
         today = pd.Timestamp.now().date()
         try:
             total_deleted = delete_published_records(conn_local, "mart.stock_profiles", today, commit=False)
-            delete_published_records(conn_supabase, "mart.stock_profiles", today, commit=False)
+            if app_env == "local":
+                delete_published_records(conn_supabase, "mart.stock_profiles", today, commit=False)
             print(f"Deleted {total_deleted} records from mart.stock_profiles for as_of_date = {today}")
             df = read_records()
             df['as_of_date'] = today
@@ -42,22 +47,25 @@ def main():
                                         ["tic", "as_of_date"], 
                                         updated_at=False, commit=False,
                                         batch_size=10)
-            insert_records(conn_supabase, df, "mart.stock_profiles", 
-                                        ["tic", "as_of_date"],
-                                        updated_at=False, commit=False,
-                                        batch_size=10)
+            if app_env == "local":
+                insert_records(conn_supabase, df, "mart.stock_profiles", 
+                                            ["tic", "as_of_date"],
+                                            updated_at=False, commit=False,
+                                            batch_size=10)
             print(f"Total records inserted = {total_inserted}")
         except Exception as e:
             print(f"Error processing earnings data: {e}")
             conn_local.rollback()
             conn_local.close()
-            conn_supabase.rollback()
-            conn_supabase.close()
+            if app_env == "local":
+                conn_supabase.rollback()
+                conn_supabase.close()
             return
         conn_local.commit()
         conn_local.close()
-        conn_supabase.commit()
-        conn_supabase.close()
+        if app_env == "local":
+            conn_supabase.commit()
+            conn_supabase.close()
         return
     return
